:truck: move extra stories to the .storybook folder and rename them
